问题修正

+++ module/tools/file_tools.py
@@ -12,7 +12,6 @@ def Compare_file_md5(file1: str, file2: str) -> bool:
             for byte_block in iter(lambda: f.read(4096), b""):
                 md5_hash.update(byte_block)
         return md5_hash.hexdigest()
-
     logger.debug(f"[Get MD5]\n"
                  f"File1:{get_file_md5(file1)}\n"
                  f"File2:{get_file_md5(file2)}")
@@ -20,11 +19,18 @@ def Compare_file_md5(file1: str, file2: str) -> bool:
 
 
 def is_used(file_name):
-    try:
-        with open(file_name, 'rb+') as f:
-            return True  # File is not locked
-    except IOError:
-        return False  # File is locked
+    if os.path.exists(file_name):
+        try:
+            with open(file_name, 'rb+') as f:
+                logger.debug(f"文件{file_name}未被占用")
+                return False  # File is not locked
+        except IOError as error:
+            logger.error(f"文件{file_name}测试失败，可能已被占用\n{error}")
+            return True  # File is locked
+    else:
+        logger.error(f"文件{file_name}不存在！")
+        return True
+
 
 def copy_file(source: str, directory: str):
     '''
@@ -56,19 +62,22 @@ def move_file(source: str, directory: str):
 	:param directory: 目标文件夹
 	:return: 成功：True 失败：False
 	'''
-    logger.debug(f"[移动]移动文件：{source} --> {directory}")
-    if not os.path.exists(source) or not os.path.exists(directory):
-        if not os.path.exists(source):
-            logger.error(f"[移动]源文件{source}不存在")
-        elif not os.path.exists(directory):
-            logger.error(f"[移动]目标文件{directory}已存在")
-        return False
-    try:
-        shutil.move(source, directory)
-        logger.debug(f"[移动]移动文件：{source} --> {directory}成功")
-        return True
-    except Exception as e:
-        logger.error(f"[移动]移动文件：{source} --> {directory}失败\n原因：{e}")
+    if not is_used(source):
+        logger.debug(f"[移动]移动文件：{source} --> {directory}")
+        if not os.path.exists(source) or not os.path.exists(directory):
+            if not os.path.exists(source):
+                logger.error(f"[移动]源文件{source}不存在")
+            elif not os.path.exists(directory):
+                logger.error(f"[移动]目标文件{directory}已存在")
+            return False
+        try:
+            shutil.move(source, directory)
+            logger.debug(f"[移动]移动文件：{source} --> {directory}成功")
+            return True
+        except Exception as e:
+            logger.error(f"[移动]移动文件：{source} --> {directory}失败\n原因：{e}")
+            return False
+    else:
         return False
 
 
@@ -79,19 +88,22 @@ def rename_file(file_path: str, new_file_name: str):
 	:param new_file_name: 新文件名
 	:return: 成功：True 失败：False
 	'''
-    logger.debug(f"[重命名]：{file_path} --> {new_file_name}")
-    if not os.path.exists(file_path) or os.path.exists(new_file_name):
-        if not os.path.exists(file_path):
-            logger.error(f"[重命名]文件{file_path}不存在")
-        elif os.path.exists(new_file_name):
-            logger.error(f"[重命名]目标文件名{new_file_name}已存在")
-        return False
-    try:
-        os.rename(file_path, new_file_name)
-        logger.debug(f"[重命名]：{file_path} --> {new_file_name}成功")
-        return True
-    except Exception as e:
-        logger.error(f"[重命名]重命名文件：{file_path} --> {new_file_name}失败\n原因：{e}")
+    if not is_used(file_path):
+        logger.debug(f"[重命名]：{file_path} --> {new_file_name}")
+        if not os.path.exists(file_path) or os.path.exists(new_file_name):
+            if not os.path.exists(file_path):
+                logger.error(f"[重命名]文件{file_path}不存在")
+            elif os.path.exists(new_file_name):
+                logger.error(f"[重命名]目标文件名{new_file_name}已存在")
+            return False
+        try:
+            os.rename(file_path, new_file_name)
+            logger.debug(f"[重命名]：{file_path} --> {new_file_name}成功")
+            return True
+        except Exception as e:
+            logger.error(f"[重命名]重命名文件：{file_path} --> {new_file_name}失败\n原因：{e}")
+            return False
+    else:
         return False
 
 
@@ -101,18 +113,22 @@ def del_file(file_path: str):
 	:param file_path: 要删除的文件
 	:return: 成功：True 失败：False
 	'''
-    logger.info(f"[删除]{file_path}")
-    if not os.path.exists(file_path):
-        logger.error(f"[删除]目标文件{file_path}不存在")
-        return False
-    else:
-        try:
-            os.remove(file_path)
-            logger.debug(f"[删除]删除文件：{file_path}成功")
-            return True
-        except Exception as e:
-            logger.error(f"[删除]文件：{file_path}删除失败\n原因：{e}")
+
+    if not is_used(file_path):
+        logger.info(f"[删除]{file_path}")
+        if not os.path.exists(file_path):
+            logger.error(f"[删除]目标文件{file_path}不存在")
             return False
+        else:
+            try:
+                os.remove(file_path)
+                logger.debug(f"[删除]删除文件：{file_path}成功")
+                return True
+            except Exception as e:
+                logger.error(f"[删除]文件：{file_path}删除失败\n原因：{e}")
+                return False
+    else:
+        return False
 
 
 def mkdir(directory_name: str):
